feat: adds owner column in HookParametersTable
