+++ app/hooks/dashboard/page.tsx
@@ -1,170 +1,14 @@
-"use client";
+import { Suspense } from "react";
+import { Skeleton } from "@chakra-ui/skeleton";
+import { fetchAddressBook } from "@/lib/data/maxis/addressBook";
+import HookParametersDashboardModule from "@/components/HookParametersDashboardModule";
 
-import React, { useState, useEffect, useMemo } from "react";
-import { useQuery } from "@apollo/client";
-import {
-  Box,
-  Container,
-  Heading,
-  Text,
-  Flex,
-  Spinner,
-  Center,
-  Alert,
-  AlertIcon,
-  AlertTitle,
-  AlertDescription,
-} from "@chakra-ui/react";
-import { HookParametersTable } from "@/components/tables/HookParametersTable";
-import { NetworkSelector } from "@/components/NetworkSelector";
-import { NETWORK_OPTIONS, networks } from "@/constants/constants";
-import {
-  GetV3PoolsWithHooksQuery,
-  GetV3PoolsWithHooksQueryVariables,
-  GetV3PoolsWithHooksDocument,
-  GqlChain,
-} from "@/lib/services/apollo/generated/graphql";
-import { Pool } from "@/types/interfaces";
-import { useSearchParams, useRouter } from "next/navigation";
-
-const HOOK_TYPE_MAP = {
-  "stable-surge": {
-    tag: "HOOKS_STABLESURGE",
-    type: "STABLE_SURGE" as const,
-    name: "StableSurge",
-  },
-  "mev-tax": {
-    tag: "HOOKS_MEVCAPTURE",
-    type: "MEV_TAX" as const,
-    name: "MEV Capture",
-  },
-} as const;
-
-type URLHookType = keyof typeof HOOK_TYPE_MAP;
-
-export default function HooksTablePage() {
-  const [selectedNetwork, setSelectedNetwork] = useState("ALL");
-  const searchParams = useSearchParams();
-  const router = useRouter();
-
-  const hookTypeParam = searchParams.get("type") as URLHookType | null;
-  const hookType =
-    hookTypeParam && HOOK_TYPE_MAP[hookTypeParam]
-      ? HOOK_TYPE_MAP[hookTypeParam]
-      : HOOK_TYPE_MAP["stable-surge"];
-
-  const { loading, error, data } = useQuery<
-    GetV3PoolsWithHooksQuery,
-    GetV3PoolsWithHooksQueryVariables
-  >(GetV3PoolsWithHooksDocument, {
-    variables:
-      selectedNetwork !== "ALL"
-        ? { chainIn: [selectedNetwork as GqlChain], tagIn: [hookType.tag] }
-        : { tagIn: [hookType.tag] },
-    context: {
-      uri:
-        selectedNetwork === "SEPOLIA"
-          ? "https://test-api-v3.balancer.fi/"
-          : "https://api-v3.balancer.fi/",
-    },
-  });
-
-  const handleNetworkChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    setSelectedNetwork(e.target.value);
-  };
-
-  const networkOptionsWithAll = useMemo(() => {
-    const baseOptions = [
-      {
-        label: "All networks",
-        apiID: "ALL",
-        chainId: "",
-      },
-    ];
-
-    // For MEV Capture hook, restrict to Base and Optimism
-    if (hookType.type === "MEV_TAX") {
-      const allowedNetworks = ["base", "optimism"];
-      return [
-        ...baseOptions,
-        ...NETWORK_OPTIONS.filter(network => allowedNetworks.includes(network.apiID.toLowerCase())),
-      ];
-    }
-
-    // For other hook types, include all networks
-    return [...baseOptions, ...NETWORK_OPTIONS];
-  }, [hookType.type]);
-
-  const networksWithAll = {
-    ...networks,
-    all: {
-      logo: "/imgs/globe.svg",
-      rpc: "",
-      explorer: "",
-      chainId: "",
-    },
-  };
-
-  useEffect(() => {
-    if (hookTypeParam && !HOOK_TYPE_MAP[hookTypeParam]) {
-      router.replace("/hooks/table?type=stable-surge");
-    }
-  }, [hookTypeParam, router]);
+export default async function HookDashboardPage() {
+  const addressBook = await fetchAddressBook();
 
   return (
-    <Container maxW="container.xl" py={8}>
-      <Flex
-        direction={{ base: "column", md: "row" }}
-        align={{ base: "flex-start", md: "center" }}
-        justify="space-between"
-        mb={6}
-        wrap="wrap"
-        gap={4}
-      >
-        <Box>
-          <Heading as="h2" size="lg" variant="special" mb={2}>
-            {hookType.name} Hook Parameters
-          </Heading>
-          <Text>View all {hookType.name} parameters for Balancer v3 pools</Text>
-        </Box>
-
-        <Box minW="250px">
-          <NetworkSelector
-            networks={networksWithAll}
-            networkOptions={networkOptionsWithAll}
-            selectedNetwork={selectedNetwork}
-            handleNetworkChange={handleNetworkChange}
-          />
-        </Box>
-      </Flex>
-
-      {loading ? (
-        <Center h="50vh" flexDir="column" gap={4}>
-          <Box p={4} w="16" h="16" rounded="full" bg="whiteAlpha.50">
-            <Spinner size="lg" color="gray.400" />
-          </Box>
-          <Text fontSize="m" color="gray.500">
-            Loading pools...
-          </Text>
-        </Center>
-      ) : error ? (
-        <Alert status="error" mt={4}>
-          <AlertIcon />
-          <AlertTitle>Error loading pools</AlertTitle>
-          <AlertDescription>{error.message}</AlertDescription>
-        </Alert>
-      ) : !data?.poolGetPools || data.poolGetPools.length === 0 ? (
-        <Alert status="info" mt={4}>
-          <AlertIcon />
-          <AlertTitle>No pools found</AlertTitle>
-          <AlertDescription>Try selecting a different network</AlertDescription>
-        </Alert>
-      ) : (
-        <HookParametersTable
-          pools={data.poolGetPools as unknown as Pool[]}
-          selectedHookType={hookType.type}
-        />
-      )}
-    </Container>
+    <Suspense fallback={<Skeleton w="full" h="500px" />}>
+      <HookParametersDashboardModule addressBook={addressBook} />
+    </Suspense>
   );
 }
